requirements.txt
pandas
numpy
scikit-learn
gdown
matplotlib
seaborn
jupyter
notebook
imbalanced-learn
nltk
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
joblib
pydantic
openai
requests
python-multipart